requirements.txt
# Core
fastapi==0.115.0
pydantic==2.7.0
pydantic-settings==2.3.0
uvicorn==0.30.0
python-multipart==0.0.9

# Database
supabase==2.5.0

# Queue / Workers
celery[redis]==5.4.0
redis==5.0.4

# HTTP
httpx==0.27.0

# Auth / JWT
PyJWT==2.9.0

# AI
anthropic==0.40.0

# Logging
loguru==0.7.2

# Real-time
sse-starlette==2.1.0

# Monitoring
flower==2.0.1

# Image processing
Pillow==10.4.0
pillow-heif==0.18.0
PyMuPDF==1.24.0

# Documents
python-docx==1.1.0
openpyxl==3.1.5

# PDF generation
WeasyPrint==62.3

# Email
resend==2.0.0
